BOL-60 refactor: Group dto 수정
