--- domain/src/main/kotlin/com/yapp/bol/group/BaseGroup.kt
@@ -1,9 +1,0 @@
-package com.yapp.bol.group
-
-open class BaseGroup(
-    val id: Long = 0,
-    val name: String,
-    val description: String,
-    val organization: String,
-    val profileImageUrl: String
-)
