+++ domain/src/main/kotlin/com/yapp/bol/group/dto/GroupMemberList.kt
@@ -1,5 +1,6 @@
-package com.yapp.bol.group
+package com.yapp.bol.group.dto
 
+import com.yapp.bol.group.Group
 import com.yapp.bol.group.member.MemberList
 
 data class GroupMemberList(
