+++ port-in/src/main/kotlin/com/yapp/bol/group/GroupService.kt
@@ -1,6 +1,8 @@
 package com.yapp.bol.group
 
 import com.yapp.bol.group.dto.CreateGroupDto
+import com.yapp.bol.group.dto.GroupMemberList
+import com.yapp.bol.group.dto.GroupWithMemberCount
 import com.yapp.bol.pageable.ApplicationSlice
 
 interface GroupService {
